Make chunked writes atomic
